rename aymurai.logging to aymurai.logger due to conflics with stdlib logging
