+++ aymurai/logger/__init__.py
@@ -1,0 +1,1 @@
+from .logger import get_logger
