+++ aymurai/models/usem/core.py
@@ -9,7 +9,7 @@ from tqdm.auto import tqdm
 import tensorflow_text  # noqa
 from more_itertools import chunked
 
-from aymurai.logging import get_logger
+from aymurai.logger import get_logger
 
 logger = get_logger(__name__)
