+++ aymurai/pipeline/postprocess.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 from tqdm.auto import tqdm
 from joblib import Parallel, delayed
 
-from aymurai.logging import get_logger
+from aymurai.logger import get_logger
 from aymurai.meta.types import DataItem, DataBlock
 from aymurai.meta.pipeline_interfaces import Transform
 from aymurai.utils.cache import cache_load, cache_save, get_cache_key
